rename model and detector
