+++ vinopy/tests/model/test_model.py
@@ -1,9 +1,9 @@
 import pytest
-from vinopy.detector.detector import Detector
-from vinopy.model.model_detect import (ModelDetectFace,
-                                       ModelDetectBody,
-                                       ModelEstimateHeadpose,
-                                       ModelEmotionRecognition)
+from vinopy.model.model import Model
+from vinopy.detector.detector_human import (DetectorFace,
+                                            DetectorBody,
+                                            # DetectorHeadpose,
+                                            DetectorEmotion)
 from PIL import Image
 import numpy as np
 # import pytest
@@ -12,7 +12,7 @@ TEST_FACE = './data/test/face.jpg'
 TEST_BODY = './data/test/person2.jpg'
 
 
-class TestDetector(object):
+class TestModel(object):
     def load_image(self, path_image=TEST_FACE):
         frame = np.array(Image.open(path_image))
         return frame
@@ -22,15 +22,15 @@ class TestDetector(object):
                                       'emotion_recognition',
                                       'estimate_headpose'])
     def test_compute(self, task):
-        detector = Detector(task)
+        model = Model(task)
         frame = self.load_image()
-        detector.predict(frame)
+        model.compute(frame)
     
     @pytest.mark.parametrize('task', ['detect_face', 
                                       'detect_body',
                                       'emotion_recognition',
                                       'estimate_headpose'])
     def test_predict(self, task):
-        detector = Detector(task)
+        model = Model(task)
         frame = self.load_image()
-        detector.predict(frame)
+        model.predict(frame)
